Clarify when we need a full render
After the hotfix 3266859a (Fix `blame` not rendering anything) we do a
fresh render whenever `regions` becomes empty which is not necessary
(and a bit confusing).  Rewrite that passage for clarity:

We need a full render whenever we have sections not currently drawn;
and also when we haven't drawn any sections at all in the previous
round.  The latter implies `not regions`.

In other words: Instead of `not regions` *or* `not current_region_names`
*and* would have been correct.

Add tests because I did this wrong 3 times now.