Fix navigation to the first section in the inline diff
For the first jumpable region, we try to show the BOF which is only
correct for views that have a header section at the top.

Inline diffs don't and so we shouldn't do anything particular as well.
Note that the historical file view (which is similar in that it shows a
whole file without headers) never did this as well.