Handle navigation in the commit panel differently
When the panel has focus  (and the graph is really visible at the top
of it), move the cursor in that ("parent") graph view.  The panel will
follow the cursor automatically.

Rverse the `[n/p]` though because the graph asks for an up/down
semantic, and `n` is in the lower register so it must mean "down", and
vice versa.