Make the toasts location flexible
Adjust `show_toast` to have flexible positions of the to be shown toast.

Allow passing in `max_width` and `location` either as integer, in which
case the value will be passed to `show_popup` without further ado.

Also allow setting floats between 0 and 1 to mean percent values of the
current viewport.  E.g. `max_width=1.0` denoting the whole viewport and
`location=0.5` denoting the middle of the screen.