Fix `pop` and `popitem` on `Cache`
If we implement `__getitem__` to call `move_to_end` on its usage, we
break `pop` and `popitem` functionality.  It seems that both call our
`__getitem__` but if the key is actually in there and to be popped,
the `move_to_end` already throws.  So basically it internally was
popped already and the internal data of the `OrderedDict` is not
coherent at this point anymore.

This can be reproduced in a REPL so race conditions due to threading
are definitely not in play here.