Limit number of commits shown in the graph
Implement limiting the graph length.  Typically draw all lines up to
the cursor (which defines what we `follow`) plus some more, defined
in `FOLLOW_UP`.

Fixes #1364
Closes #494

By doing so going to the graph can just work as long we set `follow`
when opening the graph.

Just searching for a commit hash, e.g. `ctrl+i` for the incremental
search, does not work with this approach.  Also we cannot jump to
arbitrary symbols as we don't draw all of them.