Fix initial cursor position of "File History" views
Since d19c3f81 we follow "HEAD" if nothing else can be specified when
opening a new graph.  In case "HEAD" is not within the graph we want
to show (which is only typical for file histories or other searches),
the cursor remained in BOF position.

Do a `gs_log_graph_navigate` after the initial, first draw in that
case.