Ensure user did not set `rebase.abbreviateCommands`
For our automatic interactive rebase commands, e.g. "Reword this
commit", we rewrite the todo list.  We expect long commands, e.g.
"pick", in the todo list but a user *could* change that in their
configuration.  Explicitly set `rebase.abbreviateCommands` to make
our code more robust.

We could use `-c rebase.abbreviateCommands` but this would show up
in the logging which is a bit ugly.  We also had to code this in as
yet another argument for `self.git`.

Instead use environment keys to achieve the same thing.  These
environment settings are read since v2.31.0 but naturally do not fail
on older versions.