Run the "Compare against..." wizard on the main thread
To reduce flicker (unwanted paints) run especially `show_branch_panel`
on the main thread.

Ideally we could fetch the available branches while waiting for user
input.  However, just doing it synchronously is still preferable as it
prevents the wizard's first menu from closing just to show up again a
split second later with a new panel (with the branch names to choose
from).

To defer in the second case (`show_single_line_input_panel`) was a
mistake in the first place as it only opens up an input box (which is
immediate and runs on the UI thread anyway.)