Let `[a]` act like a toggle when branches are set
Fixes #1775

We had a mismatch between the git command we executed and the prelude
we show.

The prelude suggested that `[a]` is a toggle, i.e. it did *hide* the
set `branches` when `all_branches` was also set.  But the actual git
command did not follow.

I think the prelude is right here so we fix the git command.