Set `default_branch_only` when forking a GitHub repo
Closes #1772

Setting `default_branch_only` is like a sparse fork.  See:
https://github.blog/changelog/2022-07-27-you-can-now-fork-a-repo-and-copy-only-the-default-branch/

We make the default "sparse" and introduce an argument to the command
so users can add a "create full fork" to the Command Palette or where
they wish to at their own.

Additionally, a new setting `sparse_fork` is added whose value is used
as a default when `default_branch_only` is `None`.