Teach the log panel to write out progress bars
Git's progress bars are implemented by performing simple writes with a
trailing `\r`. We check whether a line ends with a carriage return and
store the position of a virtual cursor for the next write operation.

This approach mirrors the behavior observed on the command line: if a
carriage return is detected, the (virtual) cursor remains at the initial
write position, which is at the beginning of line (bol) position.
Otherwise, the cursor moves along with the write operation and
effectively stays at the end of file (eof) position.