Implement progress updates and better timeout
We used to read line by line (`readline`) but that doesn't yield often
enough even when "--progress" is on.  That leads to two problems:
a) we might timeout as our code doesn't see any progress, b) the user
sees a blank or not-changing panel.

So we read byte by byte and form lines from it in Python land.

Note that we change the semantic of our `timeout` here.  We used to
take it as the whole time the process gets to finish, and now just
measure the time between bytes.  That is to say, we only time out
stalled processes now.