Make the "Compare x with y" action simpler to use
Fixes #1778

If the cursor is *not* on the HEAD, just offer to compare the commit
under cursor with HEAD.  (T.i. we don't show the wizard anymore.)

If the cursor *is* on the HEAD commit, show the wizard to let the user
choose a target. Make sure the wizard does not offer "HEAD" as a target
as that would mean compare "HEAD vs HEAD" which is nonsense.