Do not add "HEAD" automatically to the `gs_compare_against` wizard
We used to add "HEAD" to the candidate list automatically but the
algorithm was rather naive and resulted in showing the hint too often.
We especially want to use the proper branch names or even commit hashes
instead of the literal "HEAD" because "HEAD" is not a stable target and
the final graph would drastically change with each checkout.  But then
the wizard basically suggested to compare "HEAD" against "HEAD" which
is confusing as it is a no-op.

Remove the "magic" and instead introduce a new option `target_hints`
which optionally takes a list of strings to present as good target
candidates to the user.

Throughout our code base we only need to change two calling sites which
actually used the feature by adding:

```
	"target_hints": ["HEAD"],
```

accordingly.

Note that this generalization also serves as a preparation for future
commits where we will compute other interesting target candidates for
the user.