Keep the cursor always visible when jumping to a diff
We no longer "overscroll" diff views.  As the displayed diffs present
only portions of a file, the "hunks", there is no direct translation of
a scroll distance, our `offset`, from the real file to its corresponding
displayed diff.

Hence, check if applying the `offset` would result in overscrolling and
hiding the cursor, and if so, show the cursor centered instead.