Pass starting point as initial prompt text when creating new branches
Unlike with remote branches, local branches that were the target of the
`[b]` shortcut, were failing to pass the branch name as the initial
text of the prompt for the new branch.

This commit was already in #1797 which got reverted in #1800 because
it threw.  Since then we taught `base_commands.WithInputHandlers` to
include arguments with their default values if the command was not
called with them, essentially fixing the underlying issue.

Still, Pavel's patch needed an `or = ""` because the default
`start_point` is `None` which is not an allowed initial text for an
input handler.  Added that.

Signed-off-by: herr kaste <herr.kaste@gmail.com>