Pass around default arguments when calling argument providers
Generally, argument providers are only available for "required"
positional arguments and compute their value if not provided when
calling the command.  E.g. they ask for something using an input handler
or a quick panel.

In contrast, keyword arguments have their defaults by definition in the
method signature.

Let's pass those static default arguments down to the provider for
easier usage.  Doing so providers do not have to check for a value
being set or missing.