Fix first switch to the staged area
In a new diff view staging something and then switching the staging area
(`[tab]`) should result in showing the top of the view, the prelude,
with the cursor on the first hunk.

This did not work and was maybe never implemented correctly.  (It may
just have worked by chance, though.)