Implement `[E]` and `[W]` rebase shortcuts when viewing patches
We have `[E]` and `[W]` implemented in the graph view to quickly
initiate a rebase.  Do this also for the "show commit" and "line
history" views.  Especially "show commit" has since then learned `[n]`
and `[p]` to walk the commit history.  To e.g. reword a commit message
seems a natural fit when reviewing the commits then.