Ensure dirty decoration is in sync with the repo state
Status updates (`gs_update_status`) run lazy ("when idle") and
`gs_log_graph_refresh` calls run often `on_activated` so the state
in the store we use to determine if the repo is dirty can be outdated
easily.

To solve this issue subscribe to state updates and re-render if the
rendered state does not match the repo state by calling out to
`gs_log_graph_refresh` again.

This is costly and can be a waste *if* the refresh just started and is
still running.  We thus keep track of the time window between starting
the refresh process and actually receiving and annotating the HEAD
commit because in that window we don't need to restart the refresh.