Fix: opening multiple diff views was not possible
In the status view, a user could select multiple files with the wish
to open multiple diff views in one go.  This was broken since the
refactoring in e3519a0a (Move closing-active-view logic up, 2.41.0).

Fix it in-place instead of reverting because the code seems to be in
the right position here.