Disable `wrap_with_force` for the diff view
Since we now flash the position in the status bar wrapping is not
confusing anymore and we just do it again.

We keep this feature now turned on only for the line history view
which is otherwise still confusing as the user never knows when the
history of the selected lines "ends".