Fix caching the branches
We unconditionally cache the result of the `for-each-ref` call as
`branches` in the store although the function is parameterized and
the user can specify which `refs` she wants.

We generally want the defaults refs, "refs/heads" and "refs/remotes",
in the cache as the status view relies on it.  Update the store as
good as possible.  For more esoteric values of `refs` just skip caching
altogether to keep it simpler.