Add `merged` argument to `get_branches`
For "git: merge" we want to show only branches that are not already
fully merged or in other words are ahead by some commits.  On the
lowest level we parameterize `get_branches` to return only these
branches as the functionality is implemented for `for-each-ref`.

Be aware that we cannot cache the output of `for-each-ref` then as it
is filtered in git land.

Using "%(ahead-behind:HEAD)" in the format would be preferable and also
fast enough to use it unconditionally but is only available since git
v2.41.0 (Jun 2023).

The behavior of `--no-merged` would be equivalent of using all branches
for which `ahead > 0`.

`--[no-]merged` has been introduced in 2.7.0 (Jan 2016) on the other
hand, thus is older than GitSavvy's minimal required version.