List local branches in sections
Similar of how the overview mode of the graph view sorts the branches,
group the branches in the branches dashboard.

First show the branches that are ahead of HEAD, then the checked out
branch, then the branches that are behind, finally branches that have
diverged (are ahead and behind). Split this last section further and
put stale branches in a separate section.

This leads to a nice representation.

* E.g. the user is on main and sees all feature branches that could be
  merged right above.

* E.g. the user is on a feature branch and sees the dependent feature
  branches up to main directly below the active branch.

The feature requires git v2.41.0 and is pluggable by the
`group_by_distance_to_head_in_branch_dashboard` setting.

Closes #788