Avoid running the main `run` on the worker thread
Since `run` reads from the view, for example which branches the user
selected, it should run on the UI thread.

Now often we delegate then to other commands (`run_command`).  These
start on the UI thread so there is really no need for any thread
hopping at all.

Jump to the worker just for the git calls that may take longer, e.g.
when we push or fetch over the internet.  Jump to a separate thread for
merging since this can callback to Sublime to edit the commit message.