Fix detecting when a user wants to delete the current branch
Since git 2.43.0 the error message changed, it now starts lowercased
but also uses "used by worktree".

This actually clarifies the usage of the error and thus we add a check
that the user actually attempted to delete the currently checked out
branch.  (In other words: that the worktree is the current `repo_path`.)
Only then offer to detach and delete.