Teach the smart copy helper to copy special branches and tags
The selector for the branches ("constant.other.git.branch.git-savvy")
was too restrictive and did not catch special branch names like "main",
"dev", and so on.  This is a regression from b01ccb0b (Mark some special
branches in the graph view), released with 2.41.0 (Jun 2023).

Add the selector for tag names. It is not clear why that was missing,
so this looks like a bug as well.