Fix `gs_line_history.from_inline_diff_view` constructor
We used `is_historical_diff` diff here to decide if the shown
`actual_line` can be used directly.  This is incorrect as it is also
true when only the `base_commit` is not None.  In that case we actually
have a dynamic diff that compares against the working dir
("<base_commit>..WORKING DIR") and we need to fix the `actual_line`
accordingly.

It is, on contraire, when `target_commit` is set, that the underlying
diff of the "inline_diff"-view already has the correct value.

Note that we also start the line history from `commit_hash` so there
is a simple match here.