Remove unused `is_historical_diff`
Note that `is_historical_diff` was also wrong as for a "historical"
diff it is enough that `target_commit` is set.  If only `base_commit`
is set we still have the comparison "<base_commit>..WORKING DIR" which
is obviously a dynamic diff that merges or combines a range of
revisions.

The notion "historical" usually refers to a static, cacheable(!)
"<commit>~1..<commit>" diff.  E.g. it matches what `show_file_at_commit`
shows with the help of the gutter marks.