Fix when inline diffs need to be refreshable
`is_interactive_diff` refers to a view for which staging and unstaging
is applicable.  Basically it captures the functionality of the inline
diff *before* we introduced the "historical" feature in d6af9155
(Implement historical "inline diff", June 2020).

But a diff is always dynamic just if `target_commit` is not set as it
then essentially compares against the working dir.  Fix that here!