Implement `--intent-to-add`
When the user wants to see a diff for an untracked file, we first add
it using `--intent-to-add` and then show the diff.  (Only do this for
`e`, t.i. show diff for the file under cursor, not for the broad `f`
action!)

After that the file is shown under the new file section "ADDED". A user
can hit `d` to revert to the previous, "untracked" state.

We track these files in the `unstaged_files` section in the data layer
as that's where they belong to, and only distinguish between them in
the viewers layer.  Note that there is a difference for us and the user
as we "discard" added files differently.

Instead of the new section ("ADDED") I also tried just to mark the files
with a leading "+" (like we mark deleted files with a leading "-"). But
I liked this less.  (We also don't mark new staged files with a "+".)