Tune diff extraction for `gs_diff_open_file_at_hunk`
We used to just select the whole views content and parse a diff out of
it.  That is not quite correct as we usually have preludes before the
actual diff comes. Especially the commit message view has plenty of
content before the diff even begins and if one line begins with "commit"
or "diff" we break the parsing.  Fix that!

Select all `git-savvy.diff` regions by scope and parse each one. Select
also all `git-savvy.commit` regions as these are used for the
`show_commit` view which includes the commit header as well.