Jump to first hunk if in the header
We often see commit views with very short patches where the user might
want to jump to the source file immediately.  Make that possible!

The patch also works if the use is within the header section of a diff
and will match the first line of the first hunk of that file.  Yeah!

If the header has no hunk, e.g. because of a file rename, then we might
still not match anything and thus `next` throws a `StopIteration`.
We might also just match a different file if that has a hunk.