Fix jumping when the cursor is in EOF position
Our views usually have a trailing line at EOF because that looks better.
That line should then count as belonging to the hunk.  (At least the
code should not throw.)

Fix that in `real_linecol_in_hunk` which takes a `row_offset` into the
selected hunk.  As that can be an arbitrary value be resilient here on
the user input and pin it to the last possible line of the given hunk.