Learn to discard " R" entries
"R"-enamed files in the unstaged section are not your standard rename!
See: https://stackoverflow.com/questions/77950918/how-to-produce-the-file-status-r-in-git
We "un-intent" basically to split the " R" up.  After that we should have
two files in the status: " D" <path_alt> and "??" <path>.  The intention
is to not have any data loss.  (At the cost of possibly having more clicks
to make.)