Secure against slow ahead-behind-calculations
The commits introducing `%(ahead-behind)` indicated that its calculation
would be very fast.  That seems generally true if the "commit-graph" is
up-to-date.

Introduce a global "slow_repo" marker in the store which can maybe used
for the graph views as well.  Now probe the speed of `for-each-ref`
and fallback it is too slow.

Also run "commit-graph" and check again.  Only then set "slow_repo".