Fix: Do not set "sparse" again when re-using a graph view
Actually, we never use the argument to the constructor, but only set
`decoration` automatically.  Since we check for slow repos only on the
first, initial render, it is important to not set "sparse" automatically
again.

Note that these days git is faster probably due to their commit-graph
optimization, so I still don't set and remember the slow-ness check per
session e.g. in the `store`.  The idea is that a commit-graph update
will occur eventually and make everything fast enough again.  We'll
see how this works out.