Fix: Ensure we still follow when `gs_log_graph_refresh` timed out
When we re-run the refresh because the initial response was too slow,
i.e. after we unset "sparse", the cursor in nowheres land but the
prelude has been written.  Because of that the subsequent run will not
count as "initial_draw" and because of the cursor being in the prelude
we will essentially not "follow".

To fix this set "assume_complete_redraw" which will force the "follow"
even when the cursor is on a different spot.