Implement `gs_diff_open_file_at_hunk` when `in_cached_mode`
Surprisingly, jumping from the staged diff to the working dir was not
implemented when the file *also* had still unstaged changes.

In that case we need to translate the line from "HEAD" to the working
dir state.  (The staged diff gets applied to HEAD so its line numbers
are in relation to "HEAD".)  This matches exactly what we do in
`gs_show_file_at_commit_open_file_on_working_dir`.