Render root nodes differently
In the graph root nodes can be very hard to see (if you have multiple
of them) when they appear on the same indentation as the surrounding
"normal" commit nodes.

Let's give them a different char, `"⌂"` for now.  The exact style
does maybe not even matter.

That we read the parents, `"%p"`, might be even interesting in the
future.  It is however good to see how easy this change is to implement.