Optimize (i.e. throttle) `on_selection_modified` side-effect
We have multiple side-effects of `on_selection_modified`.  Here we
read and set `follow`.  The change is two-fold.

1. Make clear that the ref we follow only depends on the content of the
current line.  See the `lru_cache(1)` wrapping `_set_symbol_to_follow`.

2. Fix a bug where we jump to an already open graph (`[g]`) to a
specific commit that is not yet present in the graph.

I.e. the user wants to see the context of a historic commit.  She uses
`g` and we immediately focus a previously opened graph view.  Now
`on_selection_modified` gets called while the cursor has not yet
jumped but is still on the line it was before.  Hence the `lru_cache`
will filter this event.

One could say `on_selection_modified` is called not by user-intent
but only as a by-product of rendering (here: expanding) the graph, and
we MUST NOT overwrite `follow` in this case.