Catch `ProcessLookupError` when trying to kill a process
Ref #1870

Processing and drawing the graph in Sublime is much slower than reading
it from git so it is even likely that the actual process already died
when we decide that we don't need it anymore.

Catch the possible `ProcessLookupError` therefore.  Note that this is
a no-op on Windows where we kill the process differently, using the
external "taskkill" program.