Correctly await git's first emitted line
Prior we awaited the first `token` but that is a mistake as it depends
on the diff between the current and next state of the graph.  The
computation of the diff might take a long time, it might even take the
"whole" time if the graph is up-to-date and nothing needs to be drawn.

But the intention here is to get a hint of the speed of the "commit
walking", the speed of git's reading.  Make that clear in the code.