Set `project_data` after cloning has finished
We set this directly after creating the window.  But with a window
Sublime also creates a first view and this in turn triggers e.g.
`on_activated` et.al. ... and so on ... at the end SublimeLinter may
kick in and tell you that `target_dir` does not exists and can't be used
as a working dir...

Just do it later to avoid these cascades.