Make it safe to use a `ref` that gets re-written during rebase
We have two case here.

For the copy action, the rebase would get a no-op, as we copy commits
to `ref` but then keep `ref` at the same position as before, basically
overwriting what we did.  Here name the olf ref `<ref>--old` for now.

Of course a user has then to take action and probably have to delete
that "--old" ref.  But it helps understanding the process.  Let's see
how that works out.

For the extract action, the rebase would be desastrous as we move the
commits to the ref, *drop* the original commits, and then place the old
`ref` at its old position.  So we basically just loose the commits we
wanted to extract.  Here placing a `<ref>--old` label at the old
position doesn't make sense as the old place is dropped and gone.  So
we just drop the old ref too.

🤞