Ensure we can delete branches fast in the dashboard
We want to enable e.g. `ddd` to delete three branches in a row.  Hence
we must stay on the main thread ("sync" or blocking) so that after the
deletion, first the view updates and only then the next key press is
processed.

Note that we have similar problem in the status view where we call
`interface.refresh_repo_status_and_render()` to have a blocking update,
which is more elegant and self-contained.  That is not done easily here
as we call out to a command `gs_delete_branch` which is commonly used.
To make this command parametric just to change its final side-effect
`util.view.refresh_gitsavvy_interfaces(self.window)` seems more work
and might not worth it.