Ensure to denote new sem-versions if there are no remote tags
The code opted out to denote anything when no remote had any tags, which
is at least true for the pristine remote repo.  E.g. we checked:

```
            if remote_tag_names and ...:
            if remote_tags and ...:
```

Remove these (late) checks and instead rely on `self.state.remote_tags`
to be "truthy" which it is as soon as we check any remote for its tags.

This "as soon" is important for the possible initial render where we
don't have any remote tags because we didn't even *start* to fetch any.
Consider that even `self.state.remotes` might not be resolved when the
caches are clean.