Switch to short commit hashes
Normalize to *short* hashes because that is what we usually show in the
UI and hence often enough can extract from the views.  It is also cheap
to compute the short hash but not vice versa.  We also don't maintain
the underlying git-database so consistency and commit identity is not
an issue as well.

Note that we did not use long hashes before, we just used whatever got.
So this commit _normalizes_ our usage.