Fix graph refreshing on every line change under some circumstances
Just searching for "...\n" isn't safe to find *our* continuation marker
as git itself uses that if the drawn graph has holes.  E.g. if we
pick-axe or limit the graph to certain files not all commits are shown.
In other words: our continuation marker is not unique.

Just extract the last few characters of the graph view to check for the
marker which should be at EOF position all the time.