Implement `query` parameter when searching for PR's
Fixes #1937

Add `query` argument to `gs_github_pull_request` (default: "") to filter
the list of PR's we fetch.  As the search API does return a subset of
the information we need, query the full PR response in a second step
(but in the background using a future).

The query parameter is a string and can be basically anything a user
would also type in on Github's website.

Interesting candidates are e.g. "involves:@me" which is also added to
the Command Palette.