Switch to short hashes (for now)
In most views we use and show short hashes, e.g. the graph.  Since we
extract these hashes and pass them along, normalize and switch to short
hashes (for now) for the purpose of the caches.

This would be a good and cheap move *if* we only ever used actual
"hashes" when we type "commit_hash".  But we might use "refs" here as
well under some circumstances which we can't shorten.

Either a) find these places where we might introduce "refs" and resolve
them or b) maintain a big mapping from short to long hashes, populated
during graph and blame parsing.