Optimize `[N]` and `[P]` after staging
When a user staged a file, GitSavvy will switch to "staged" mode as soon
as the diff gets clean.  This is implemented, a) to actually see what's
gonna be committed and b) e.g. to issue "Line History" directly to
select a commit to fixup.

Unfortunately this broke `gs_diff_switch_files` intuitive UX.  As we
switched to the staged mode, a subsequent `N` could not select a useful
*next* file to continue staging.

This commit fixes that.  The intended UX is, e.g.
    `[S]` to stage a file, followed by
    `[N]` to switch to the next _unstaged_ file.