Add setting `fetch_new_remotes: true`
When adding remotes, we always asked if we should now also fetch this
remote.  Add "fetch_new_remotes" and just fetch without asking if set
to `True`.

This is also the default because frankly you likely want that.  (Except
for huge repos, but then set this in the project settings, no?)