Remove special "author" handling
Probably because of fear, we still had special handling for the "author"
constraint, e.g. `filter_by_author`.

(a) But when set, the prelude didn't show it.
(b) The edit filter feature explicitly has the author filtering in its
help.

Just remove that and use `filters`.  This way,we have it in the prelude,
can edit it with `f` and toggle it with `F`.  Users can still access it
from outside as "filters" is exposed in the `gs_graph` constructor.

Win-win?  No compatibility mode included, users must adjust their custom
commands.